Safari-609.1.11.1/LayoutTests/crypto/workers/subtle/ec-generate-export-raw-key-expected.txt

[Worker] Test exporting an ECDH public key with raw format in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ec-generate-export-raw-key.js
[Worker] Generating a key pair...
[Worker] Exporting the public key...
PASS [Worker] publicKey.byteLength is 65
PASS successfullyParsed is true

TEST COMPLETE
